// src/features/transportes/types/transporte.types.ts
export type Transporte = {
  id: number;
  parcelaId: string;
  parcelaNombre?: string;
  camionMatricula: string;
  camionMarca?: string;
  camioneroDni: string;
  camioneroNombre?: string;
  fechaCarga: string;
  toneladasCargadas: number;
  tipoMadera: string;
  destino: string;
  centroId?: number;
  centroNombre?: string;
  eudrCumplimiento: boolean;
  eudrGeolocalizacionOk: boolean;
  eudrReferenciaCatastralOk: boolean;
  pefcCumplimiento: boolean;
  sureCumplimiento: boolean;
  parcelaReferenciaCatastral?: string;
  parcelaWktGeolocalizacion?: string;
  razonesIncumplimiento?: string[];
  codigoTrazabilidad: string;
  fechaCreacion?: string;
  nombre?: string;
  ddsOk?: boolean | null;
  referenciaDds?: string | null;
  cocPefcOk?: boolean;
  referenciaCocPefc?: string | null;
  cocSureOk?: boolean;
  geiReduccionOk?: boolean | null;
  referenciaCocSure?: string | null;
};

export type TransporteFormData = {
  nombre?: string;
  parcelaId: string;
  camionMatricula: string;
  camioneroDni: string;
  fechaCarga: string;
  toneladasCargadas: number;
  tipoMadera: string;
  destino: string;
  centroId?: number;
  parcelaReferenciaCatastral?: string;
  ddsOk?: boolean | null;
  referenciaDds?: string | null;
  cocPefcOk?: boolean;
  referenciaCocPefc?: string | null;
  cocSureOk?: boolean;
  geiReduccionOk?: boolean | null;
  referenciaCocSure?: string | null;
};